Register app stuff via Application class
